requirements.txt
Pillow==5.4.1
pytesseract==0.2.6
pytest==4.3.1
coverage==4.5.3
pytest-cov==2.6.1
StringDist==1.0.9
